% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cl.R
\name{cl}
\alias{cl}
\alias{command_line}
\alias{is.command_line}
\alias{print.command_line}
\title{Vector of the full command line as seen by R}
\usage{
cl(x, ...)

command_line(x = commandArgs())

is.command_line(x)

\method{print}{command_line}(x, ...)
}
\arguments{
\item{x}{vector of command line arguments. (Default: \code{commandArgs()})

\code{cl} and \code{cl_cl} return the full, unprocessed command-line vector.
There is no differnce from \code{commandArgs} except for allowing for an
overwrite of \code{commandArgs}; this is used mainly for testing purposes.

These are non-exported functions.}
}
\value{
character contained all arguments accessible by R.
}
\description{
Vector of the full command line as seen by R
}
\seealso{
\code{\link[base]{commandArgs}}
}

