% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_args.R
\name{command_args}
\alias{command_args}
\alias{command_args.NULL}
\alias{command_args.character}
\title{Command Arguments}
\usage{
command_args(x = NULL, trailingOnly = FALSE)

\method{command_args}{NULL}(x = NULL, trailingOnly = FALSE)

\method{command_args}{character}(x, trailingOnly = FALSE)
}
\arguments{
\item{x}{command_line object or character;}
}
\value{
character; everything after the command verb; it includes options and targets.
}
\description{
Get the command arguments (even) if the command is not present.
}
\details{
Verb arguments are everthing after the verb -- if it exists.

command_args operates is based on 'arg'
}
\examples{
  args1 <- c( "Rscript", "-a", "-b", "--args", "verb", "-c", 3, "-d" ) 
  
  args1 \%>\% command_args 
  args1 \%>\% command_args( trailingOnly = TRUE )
  
  # Options missing
  args2 <- c( "Rscript", "-a", "-b", "--args", "verb", "target1", "target2" ) 

  # 
  args3 <- c( "Rscript", "-a", "-b")
  args3 \%>\% command_args                # character(0) cl_args is char(0)
}

