% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_line.R
\name{command_line}
\alias{command_line}
\title{Command line}
\usage{
command_line(x = command_args())
}
\arguments{
\item{x}{vector of command line arguments usually \code{commandArgs()}}
}
\value{
character vector of class 'command_line' containing all arguments 
accessibles to \code{command_args} / \code{commandArgs}
}
\description{
Vector of the full command line as seen by \code{commandArgs}.
}
\details{
\code{command_line} return the full, unprocessed command-line vector. 
There is no differnce from \code{command_args} except the return value is 
of class 'command_line'.
}
\examples{

  command_line() 
  
  args <- c("Rscript", "-a", "-b", "--args", "verb", "-c", "3", "-d")
  command_line(args)
  
}
\seealso{
\code{\link[base]{command_args}} \cr
  \code{\link[base]{commandArgs}}
}

