% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exe.R, R/exe_args.R
\name{exe}
\alias{exe}
\alias{exe_args}
\title{Get the executbale}
\usage{
exe(x)

exe_args(x)
}
\arguments{
\item{x}{character or command_line class; (Default: \code{cl()} )}

\item{x}{object from which to deermine exe arguments. (Default:cl())

The executable arguments are those that follow the interpreter.

Typically is}
}
\value{
vector inheriting from 'exe_args', 'command_line', 'character'
}
\description{
Returns the executable as found in the commandArgs

Get the executable arguments
}
\details{
Returns the executable as found from the command line
}
\examples{
exe() # Assumes cl
   exe( commandArgs() ) # same as above
exe_args()  # assumes current application's arguments
  exe_args( c("--args", "verb", "--foo", "bar" ) )
}
\seealso{
\code{\link{exe_args}} for the arguments to the executable including the
  name of the currently executing program
}

