% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe_method.R
\name{exe_method}
\alias{exe_method}
\title{exe_method}
\usage{
exe_method()
}
\value{
character, one of 'rstudio', 'console', 'batch' or 'rscript'
}
\description{
Determine the execution method of the current running programm
}
\details{
Determines how the current program was invoked whether within Rstudio
}

