% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe.R
\name{exe_name}
\alias{exe_name}
\alias{exe_opts}
\alias{interpreter_name}
\alias{interpreter_opts}
\title{Get the executbale (interpreter) name and options}
\usage{
exe_name(x = command_args())

interpreter_name(x = command_args())

exe_opts(x)

interpreter_opts(x)
}
\arguments{
\item{x}{character or command_line; The vector from which to parse the 
executable (Default: \code{command_line()} )}
}
\value{
string path to executable.
}
\description{
Returns the executable (interpreter) as found in the command_args/commandArgs
}
\details{
Returns the executable as found from the command line.
}
\examples{
   exe_name()                 # assumes command_args() 
   exe_name( command_args() ) # same as above 
   exe_name( NULL )
   
   exe_opts()  # assumes current application's arguments
   exe_opts( c( "R", "--vanilla", "--args", "verb", "--foo", "bar" ) )
   
}
\seealso{
\code{\link{exe_opts}} for the arguments to the executable including the 
  name of the currently executing program
}

