% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exe.R, R/exe_opts.R
\name{exe}
\alias{exe}
\alias{exe_opts}
\title{Get the executbale (interpreter)}
\usage{
exe(x)

exe_opts(x)
}
\arguments{
\item{x}{character or command_line; The vector from which to parse the 
executable (Default: \code{command_line()} )}

\item{x}{object from which to deermine exe arguments. 
         (Default: \code{command_line()} )

The executables options are all arguments following the executable path up 
until the \code{--args} flag.}
}
\value{
string path to executable.

vector inheriting from 'exe_opts', 'command_line', 'character'
}
\description{
Returns the executable (interpreter) as found in the command_args/commandArgs

Get the executable options.
}
\details{
Returns the executable as found from the command line.
}
\examples{
   exe() # Assumes command_line
   exe( commandArgs() ) # same as above 
   exe( NULL )
   
  exe_opts()  # assumes current application's arguments
  exe_opts( c("--args", "verb", "--foo", "bar" ) )

}
\seealso{
\code{\link{exe_opts}} for the arguments to the executable including the 
  name of the currently executing program
}

