% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expand_opts}
\alias{expand_opts}
\title{Expand argument vector to split names from values}
\usage{
expand_opts(opts = commandArgs())
}
\arguments{
\item{opts}{character vector of arguments. (Default: commandArgs())}
}
\description{
Expand argument vector to split names from values
}
\details{
This is an internal function and should generally not be called
directly.

\code{expand_opts} does two things:
\itemize{
  \item{"1."}{ removes values preceding and including --args }
  \item{"2."}{ splits and value containing an equal (=) sign }
}
}
\examples{
optigrab:::expand_opts()
optigrab:::expand_opts( opts=c( "--foo", "bar") )
}
\seealso{
\code{\link{grab_opt}}
}
\keyword{manip}

