% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gnu_style.R
\docType{data}
\name{gnu_style}
\alias{gnu_style}
\title{GNU-style command line options}
\format{\preformatted{List of 3
 $ flag_test   :function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 4 3 4 79 3 79 4 4
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xbcc11c8> 
 $ flag_to_name:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 6 21 6 52 21 52 6 6
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xbcc11c8> 
 $ name_to_flag:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 8 21 15 1 21 1 8 15
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xbcc11c8> 
}}
\usage{
gnu_style
}
\description{
Functions for enabling GNU-style commpand-line option behavior.
}
\details{
Functions for enabling GNU-style command-line options. GNU-style options
  are characterized by a single dash (\code{-}) before single character
  option flags and a double dash (\code{--}) before multiple character
  option flags.

  By convention, gnu style options flags must begin with a letter; if numbers
  were allowed option flags would be ambiguous with negative option values.
}
\references{
\url{http://www.gnu.org/prep/standards/standards.html}
}
\seealso{
Non-exported function \code{*_flag_test}, \code{*_flag_to_name} and
  \code{*_name_to_flag} \cr
  \code{\link{gnu_style}} \cr
  \code{\link{java_style}} \cr
  \code{\link{ms_style}}
}
\keyword{datasets}

