% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_style.R
\docType{data}
\name{java_style}
\alias{java_style}
\title{Java-style command line options}
\format{An object of class \code{list} of length 3.}
\usage{
java_style
}
\description{
Functions for enabling Java-style commpand-line option behavior.
}
\details{
Functions for enabling Java-style command-line options. Java-style options
are characterized by a single dash (\code{-}) before the option name.

By conventions, Java-style options cannot must begin with a upper or lower
case letter.
}
\seealso{
\itemize{
\item Non-exported function \code{*_flag_test}, \code{*_flag_to_name} and
\code{*_name_to_flag}
\item \link{gnu_style}
\item \link{java_style}
\item \link{ms_style}
}
}
\keyword{datasets}
