% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ms_style.R
\docType{data}
\name{ms_style}
\alias{ms_style}
\title{Microsoft-style command line options}
\format{\preformatted{List of 3
 $ flag_test   :function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 3 17 3 49 17 49 3 3
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7312ef8> 
 $ flag_to_name:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 5 20 5 49 20 49 5 5
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7312ef8> 
 $ name_to_flag:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 7 20 7 49 20 49 7 7
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7312ef8> 
}}
\usage{
ms_style
}
\description{
Functions for enabling Microsoft-style commpand-line option behavior.
}
\details{
Functions for enabling Microsoft-style command-line options.
  Microsoft-style options are characterized by a single forward slash
  (\code{/}) before the option name.

  Microsoft-style options can be supported by seetting
}
\examples{
opt_style(ms_style)
  opt_get( "foo", opts=c("/foo", "bar") )
}
\seealso{
Non-exported function \code{*_flag_test}, \code{*_flag_to_name} and
  \code{*_name_to_flag} \cr
  \code{\link{gnu_style}} \cr
  \code{\link{ms_style}} \cr
  \code{\link{ms_style}}
}
\keyword{datasets}

