% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/opt_assign.R
\name{opt_assign}
\alias{opt_assign}
\title{Get option and assign to variables}
\usage{
opt_assign(x, pos = 1, inherits = FALSE, flag = x, ...)
}
\arguments{
\item{x}{a variable name, given as a character string. No coercion is done, and the first element of a character vector of length greater than one will be used, with a warning.}

\item{pos}{where to do the assignment. By default, assigns into the current environment. See ‘Details’ for other possibilities.}

\item{inherits}{should the enclosing frames of the environment be inspected?}

\item{flag}{character; name of option}

\item{...}{arguments passed to \code{opt_get}

If}
}
\description{
opt_get + sssign
}
\examples{
opt_assign( "foo", opts=c("--foo","bar") )
}
\seealso{
\code{\link{opt_get}} \cr
  \code{\link[base]{assign}}
}

