% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opt_get_args.R
\name{opt_get_args}
\alias{opt_get_args}
\title{Return arguments to Rscript}
\usage{
opt_get_args(opts = commandArgs())
}
\arguments{
\item{opts}{character; vector of arguments, (Default: \code{commandArgs()})

Returns the user provided arguments, i.e. those following (the first)
\code{--args} flag.}
}
\value{
character; the \code{commandArgs()} vector stripping values preceding and
  including (the first) \code{--args} flag.
}
\description{
Arguments to Rscript are those up to the \code{--args} argument.
}
\examples{
opt_get_args()
  opt_get_args( opts=c( "--foo", "bar") )
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

