% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/opt_get_command.R
\name{opt_get_command}
\alias{opt_get_command}
\title{Command to the Rscript.}
\usage{
opt_get_command(opts = commandArgs())
}
\arguments{
\item{opts}{character; Vector from which to parse options
  (default: \code{commandArgs()} )}
}
\value{
character of length 1; the command
}
\description{
Return the command of the Rscript, generally the first argument following
#' \code{--args}.
}
\note{
this assumes that any flags occurring before the command has
exactly 1.  value. A command such as "> myscript --verbose subcmd" will be
misparsed; the code will assume that "subcmd" is the value of the
flag "--verbose"
}
\examples{
opt_get_command()
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

