% Generated by roxygen2 (4.0.2): do not edit by hand
\name{opt_get_path}
\alias{opt_get_path}
\title{Grab the path to the Rscript.}
\usage{
opt_get_path(opts = commandArgs())
}
\arguments{
\item{opts}{character; Vector from which to parse options (default: \code{commandArgs()} )}
}
\value{
character; path to Rscript
}
\description{
Grab the path to the Rscript.
}
\note{
non-exported
}
\examples{
optigrab:::opt_get_path()
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

