% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/opt_grab.R
\name{opt_grab}
\alias{opt_get}
\alias{opt_grab}
\title{Get option's values from the command-line}
\usage{
opt_grab(flag, default = NA, n = 1, required = FALSE,
  description = NULL, opts = commandArgs())

opt_get(flag, default = NA, n = 1, required = FALSE, description = NULL,
  opts = commandArgs())
}
\arguments{
\item{flag}{character vector.  The flags that may be associated with the that denote the command line opts}

\item{default}{the value should the value not be provided}

\item{n}{(integer) number of values to retrieve (default: 1)}

\item{required}{(logical) whether the value is required.}

\item{description}{(character) message to be printed with \code{opt_help}}

\item{opts}{(character) vector to parse for options}
}
\value{
a value parsed the opts vector associated with the flag.
}
\description{
Returns value(s) from the command-line associated with the desired option.
}
\details{
opt_grab parses the command line vector extracting value(s) that are
identified by one or more \code{flag}s.

command line"flags.   It is useful with code executed using \code{Rscript},
a \code{#!} on linux or \code{R CMD BATCH}. By default, it closely follows
the ubiquitous GNU standards for command-line interfaces. It departs from
other command-line interface packages by avoiding complex specifications in
favor of a lighter more straight-forward interface.

\code{flag} (charater) vector of possible synonyms identifying the values for
This can be one or more possible flags allowing for synonyms or values.

\code{default} provides value(s) for the option if no value is found.

\code{n} the number of value(s) to retrieve from the command line.  If
\code{n=0}, then a logical value is returned indicating whether the flag
exists

\code{required} indicates if a value is required. If the the flag is not
found or if there is not the correct number of value(s) are not found. If
\code{default} is given, then no missing value error will occur.

\code{help} is a short message about the option. It is concatenated with the
\code{flag} option and stored in the 'optigrab' option.  This is used with
\code{\link{opt_help}}.

\code{opts} is the vector from which options are parsed. By default, this is
\code{commandArgs()}.
}
\examples{
opts <- c( '--foo', 'bar' )
  optigrab::opt_grab( c('--foo') )
  optigrab::opt_grab( c('--foo'), opts=opts )
}
\references{
http://www.gnu.org/prep/standards/standards.html
}
\seealso{
\code{\link{commandArgs}}
}
\keyword{utils}

