% Generated by roxygen2 (4.0.2): do not edit by hand
\name{opt_help}
\alias{opt_help}
\title{Provide program usage information and exit.}
\usage{
opt_help(opts = commandArgs())
}
\arguments{
\item{opts}{character.  Vector from which to parse options (default: \code{commandArgs()} )}
}
\description{
Prints program usage information and exits. Usage information comes from
the \code{help} arguments to \code{opt_grab}. Calls to
\code{opt_grab} following the \code{opt_help} call will not be considered, so
it is best to specify make all \code{opt_grab} calls before \code{opt_help}.
}
\examples{
opts <- c( "--foo", "bar")
optigrab:::opt_grab( "--foo", description="Specifies Foo" )
optigrab:::opt_help()
}
\seealso{
\code{\link[base]{commandArgs}}
}
\keyword{utils}

