% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_names.R
\name{opt_names}
\alias{opt_names}
\title{Option Names}
\usage{
opt_names(opts = command_args(), style = getOption("optigrab")$style)
}
\arguments{
\item{opts}{character; vector of command arguments}

\item{style}{list describing style elements}
}
\description{
Get option names from the command-line
}
\examples{
   opts <- c("--foo", "env-fill", "--bar", "555")
  
   flags(opts)     # --foo, --bar
   opt_names(opts) # foo, bar
   
   opts <- c("-foo", "env-fill", "-bar", "555")
   
   flags(opts, java_style)     # --foo, --bar
   opt_names(opts, java_style) # foo, bar
   
}
\seealso{
\code{\link{flags}}
}

