% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opt_split_args.R
\name{opt_split_args}
\alias{opt_split_args}
\title{Split command arguments}
\usage{
opt_split_args(opts = commandArgs())
}
\arguments{
\item{opts}{character; vector of arguments. (Default: commandArgs())

\code{opt_split_args} splits and value containing an equal (=) sign}
}
\description{
Splits command argument vector to name, value pairs.
This is an internal function and should generally not be called
directly.
}
\note{
non-exported
}
\examples{
optigrab:::opt_split_args()
  optigrab:::opt_split_args( opts=c( "--foo=hello", "-b=goodbye") )
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

