% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{script}
\alias{script}
\alias{script.NULL}
\alias{script.character}
\alias{script.default}
\title{script}
\usage{
script(x = command_args(), ...)

\method{script}{character}(x)

\method{script}{default}(x)

\method{script}{NULL}(x)
}
\arguments{
\item{x}{command_line or character;}
}
\description{
Get the name of the script as it appears on the command line.
}
\details{
Returns the script if running as a Rscript.
}
\examples{
  cl <- c("/opt/R/R-3.2.1/lib/R/bin/exec/R", "--slave", "--no-restore", 
  "--vanilla", "--file=./r-command-args.r", "--args", "--flag=value" )

  script(cl)
  script(cl) \%>\% normalizePath
  
  script()
  
}

