% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{script_name}
\alias{script_name}
\title{Get running script or script options}
\usage{
script_name(x = command_args())
}
\arguments{
\item{x}{character vector representing the command line.}
}
\description{
Get the name of the script as it appears on the command line.
}
\details{
Returns the script if running as a Rscript.
}
\examples{
  cl <- c("/opt/R/R-3.2.1/lib/R/bin/exec/R", "--slave", "--no-restore", 
  "--vanilla", "--file=./r-command-args.r", "--args", "--flag=value" )

  script()    # NULL
  script(cl)  # "./r-command-args.r"
  # script(cl) \%>\% normalizePath
  
  script_opts() # NULL
  script_opts(cl)
  
}

