% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_opts.R
\name{script_opts}
\alias{script_opts}
\alias{script_opts.command_line}
\title{script_opts}
\usage{
script_opts(x)

\method{script_opts}{command_line}(x)
}
\arguments{
\item{x}{object}
}
\description{
Get the script options
}

