% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_on_eq.R
\name{split_on_eq}
\alias{split_on_eq}
\title{Split Elements on Equal}
\usage{
split_on_eq(x)
}
\arguments{
\item{x}{character; vector of values.

\code{split_on_eq} splits each element of \code{x} on the equal (=) 
sign, expanding the array such that \code{"--foo=bar"} becomes 
\code{c("--foo", "bar")}.

Will currently, split does not ignore `=` in quotes.}
}
\description{
Splits vector elements on '=' inserting additional elements at the point of 
the split. This is an internal utility and should generally not be called 
directly.
}
\note{
non-exported
}
\examples{
  \dontrun{
    split_on_eq( c( "--foo=bar", "-b=qux") )       # --foo, bar -b qux
    split_on_eq( c( "--foo 'b=r'", "-b=goodbye") ) # fails
  }
}
\seealso{
\code{\link{opt_grab}} \cr
}

