% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_to_cl.R
\name{str_to_cl}
\alias{str_to_cl}
\alias{str_to_opts}
\title{Convert a string to command_line vector}
\usage{
str_to_cl(x = character())

str_to_opts(...)
}
\arguments{
\item{x}{character; string to parse as if it is a command line

This is an internal function used predominantly for testing. It might be
deprecated in the near future.}
}
\value{
A character array that could be similar to that  provided by
  \code{commandArgs}.
}
\description{
Split a string based on whitespace ignore single- and double quoted entries
}
\note{
not-exported, by design
}
\examples{
\dontrun{
    str <- 'cmd -t "Say Anything" --character \\'Lloyd Dobler\\''
    str_to_cl(str)
    split_ws_nonquote(str)
  }
}
\seealso{
\code{\link[base]{commandArgs}} \cr
  \code{\link{cl}}
}

