% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/str_to_opts.R
\name{str_to_opts}
\alias{str_to_opts}
\title{Split a string bases on whitespace}
\usage{
str_to_opts(x = character())
}
\arguments{
\item{x}{character. String to split into a command line}
}
\description{
Split a string based on white to more easily get
}
\details{
This is an internal function used predominantly for testing. It might be
deprecated in the near future.
}
\note{
not-exported
}
\seealso{
\code{\link[base]{commandArgs}}
}

