% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.logical.character}
\alias{as.logical.character}
\title{Transforms a character vector to logical}
\usage{
\method{as.logical}{character}(x, ...)
}
\arguments{
\item{x}{vector to be transformed into}

\item{...}{additional arguments}
}
\description{
Transforms character vector to logical vector.  0's and 1's are transformed
to TRUE and FALSE respectively.  All other values result in NAs.
}
\examples{
optigrab:::as.logical.character( 0:2 )  # FALSE, TRUE, NA
}

