% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{flag_test_gnu}
\alias{flag_name_parser_gnu}
\alias{flag_test_gnu}
\alias{gnu_style}
\title{GNU style command line parsing functions}
\format{\preformatted{List of 2
 $ flag_test       :function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 10 18 10 57 18 57 10 10
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x5245270> 
 $ flag_name_parser:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 14 25 14 56 25 56 14 14
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x5245270> 
}}
\usage{
flag_test_gnu(x)

flag_name_parser_gnu(x)

gnu_style
}
\arguments{
\item{x}{Character of options}
}
\value{
logical indicationg which arguments are gnu style flags

character.  The names of the flags
}
\description{
Internal functions that are used to specify the parsing style.
}
\keyword{datasets}
\keyword{utils}

