% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gnu_style.R
\docType{data}
\name{flag_test_gnu}
\alias{flag_name_parser_gnu}
\alias{flag_test_gnu}
\alias{gnu_style}
\title{GNU style command line parsing functions}
\format{\preformatted{List of 2
 $ flag_test       :function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 11 20 11 59 20 59 11 11
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x566bc98> 
 $ flag_name_parser:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 17 27 17 58 27 58 17 17
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x566bc98> 
}}
\usage{
flag_test_gnu(x)

flag_name_parser_gnu(x)

gnu_style
}
\arguments{
\item{x}{Character of options}
}
\value{
logical indicationg which arguments are gnu style flags

character.  The names of the flags
}
\description{
Internal functions that are used to specify the parsing style.
}
\keyword{datasets}
\keyword{utils}

