% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnu_style.R
\docType{data}
\name{gnu_style}
\alias{gnu_style}
\title{GNU-style command line options}
\format{An object of class \code{list} of length 3.}
\usage{
gnu_style
}
\description{
Functions for enabling GNU-style commpand-line option behavior.
}
\details{
Functions for enabling GNU-style command-line options. GNU-style options
are characterized by a single dash (\code{-}) before single character
option flags and a double dash (\code{--}) before multiple character
option flags.

By convention, gnu style options flags must begin with a letter; if numbers
were allowed option flags would be ambiguous with negative option values.
}
\references{
[GNU Command Line Standards]{http://www.gnu.org/prep/standards/standards.html}
}
\seealso{
\itemize{
\item Non-exported function \code{*_flag_test}, \code{*_flag_to_name} and
\code{*_name_to_flag}
\item \link{gnu_style}
\item \link{java_style}
\item \link{ms_style}
}
}
\keyword{datasets}
