% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/java_style.R
\docType{data}
\name{java_style}
\alias{java_style}
\title{Java-style command line options}
\format{\preformatted{List of 3
 $ flag_test   :function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 3 19 3 51 19 51 3 3
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x733e1a8> 
 $ flag_to_name:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 5 22 5 51 22 51 5 5
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x733e1a8> 
 $ name_to_flag:function (x)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 7 22 7 51 22 51 7 7
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x733e1a8> 
}}
\usage{
java_style
}
\description{
Functions for enabling Java-style commpand-line option behavior.
}
\details{
Functions for enabling Java-style command-line options. Java-style options
  are characterized by a single dash (\code{-}) before the option name.
}
\seealso{
Non-exported function \code{*_flag_test}, \code{*_flag_to_name} and
  \code{*_name_to_flag} \cr
  \code{\link{gnu_style}} \cr
  \code{\link{java_style}} \cr
  \code{\link{ms_style}}
}
\keyword{datasets}

