% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opt_assign.R
\name{opt_assign}
\alias{opt_assign}
\alias{opt_assign_all}
\title{Get option and assign to variables}
\usage{
opt_assign(x, pos = 1, inherits = FALSE, name = x, ...,
  assign.na = FALSE)

opt_assign_all(x, ..., assign.na = FALSE)
}
\arguments{
\item{x}{character or list; variable names or alias. If no coercion is done,
and the first element of a character vector of length greater than
one will be used, with a warning.}

\item{pos}{where to do the assignment. By default, assigns into the current
environment. See 'Details' for other possibilities.}

\item{inherits}{should the enclosing frames of the environment be inspected?}

\item{name}{character; name of option}

\item{...}{arguments passed to \code{opt_get}}

\item{assign.na}{logical; whether \code{NA} can be assigned.
       (Default: \code{FALSE})

\code{opt_assign} parses an option using \code{opt_get} and then assigns it
according to ...}
}
\description{
opt_get and assign
}
\examples{
opt_assign( "foo", opts=c("--foo","bar") )
  opt_assign( c("foo","f"), opts=c("--foo","baz") )
}
\seealso{
\code{\link{opt_get}} \cr
  \code{\link[base]{assign}}
}

