% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_expand.R
\name{opt_expand}
\alias{opt_expand}
\title{Expand opts to full vector}
\usage{
opt_expand(x = command_line(), style = getOption("optigrab")$style)
}
\arguments{
\item{opts}{character vector of arguments. (Default: cl())}
}
\description{
Expand option vector to split names from values. This is an internal function 
and should generally not be called directly.
}
\details{
\code{opt_expand} does two things: 
\itemize{
  \item{"1."}{ removes values preceding and including --args }
  \item{"2."}{ splits and value containing an equal (=) sign }
}
}
\note{
non-exported
}
\examples{
  opt_expand()
  optigrab:::opt_expand( cl=c( "--foo", "bar") ) 
  optigrab:::opt_expand( cl=c( "--foo=bar") ) 
  
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

