% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_expand.R
\name{opt_expand}
\alias{opt_expand}
\title{Expand opts to full vector}
\usage{
opt_expand(opts = commandArgs())
}
\arguments{
\item{opts}{character vector of arguments. (Default: \code{\link[base:commandArgs]{base::commandArgs()}} )}
}
\description{
Expand option vector to split names from values. This is an internal function
and should generally not be called directly.
}
\details{
\code{opt_expand} does two things:
\itemize{
#1. Removes values preceding and including \code{--args}.
#2. Splits and value containing an equal (`=``) sign.
}
}
\note{
non-exported
}
\examples{
  optigrab:::opt_expand()
  optigrab:::opt_expand( opts=c( "--foo", "bar") ) 
  
}
\seealso{
\itemize{
\item \code{\link[=opt_grab]{opt_grab()}}
\item \code{\link[base:commandArgs]{base::commandArgs()}}
}
}
