% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opt_expand.R
\name{opt_expand}
\alias{opt_expand}
\title{Expand argument vector to split names from values
This is an internal function and should generally not be called
directly.}
\usage{
opt_expand(opts = commandArgs())
}
\arguments{
\item{opts}{character vector of arguments. (Default: commandArgs())

\code{opt_expand} does two things:
\itemize{
  \item{"1."}{ removes values preceding and including --args }
  \item{"2."}{ splits and value containing an equal (=) sign }
}}
}
\description{
Expand argument vector to split names from values
This is an internal function and should generally not be called
directly.
}
\note{
non-exported
}
\examples{
optigrab:::opt_expand()
  optigrab:::opt_expand( opts=c( "--foo", "bar") )
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

