% Generated by roxygen2 (4.0.2): do not edit by hand
\name{opt_get_args}
\alias{opt_get_args}
\title{Return actual arguments to R script
This is an internal function and should generally not be called
directly.}
\usage{
opt_get_args(opts = commandArgs())
}
\arguments{
\item{opts}{character vector of arguments. (Default: commandArgs())

\code{opt_get_args} removes values preceding and including --args}
}
\description{
Return actual arguments to R script
This is an internal function and should generally not be called
directly.
}
\note{
non-exported
}
\examples{
optigrab:::opt_get_args()
  optigrab:::opt_get_args( opts=c( "--foo", "bar") )
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

