% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opt_get_command.R
\name{opt_get_command}
\alias{opt_get_command}
\title{Command to the Rscript.}
\usage{
opt_get_command(opts = commandArgs())
}
\arguments{
\item{opts}{character; Vector from which to parse options
  (default: \code{commandArgs()} )}
}
\value{
character of length 1; the command found from the command-line.
\code{NA} if a command cannot be identified.
}
\description{
Return the command for the application. The command is the
  * first argument not accounted for by the flags specification
}
\details{
\code{opt_get_command} look for the first unaccounted for options (after
\code{--args}). The number of options needed by each flags is determined by
and saved by calls to \code{\link{opt_get}}. See the details to see how
\code{n} is automatically determined.  When not explicitly defined the number
of options needed by each flag is 1.  Becasue of this, it customary to call
\code{opt_get_command} after all \code{opt_get} commands.  For most simple
applications, it likely doesn't matter.

assumes any flags occurring before the command has
exactly 1 value. A command such as "> myscript --verbose command" will be
misparsed; the code will assume that "command" is the value of the
flag "--verbose"
}
\examples{
opt_get_command()
}
\seealso{
\code{\link{opt_get}} \cr
  \code{\link{base}{commandArgs}}
}

