% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opt_get_command.R
\name{opt_get_command}
\alias{opt_get_command}
\title{Command to the Rscript.}
\usage{
opt_get_command(opts = commandArgs())
}
\arguments{
\item{opts}{character; Vector from which to parse options
  (default: \code{commandArgs()} )}
}
\value{
character of length 1; the command
}
\description{
Return the command for the application. The command is the
  * first argument not accounted for by the flags specification or,
  * last non-flag commandline argument.
}
\note{
\code{opt_get_command} assumes any flags occurring before the command has
exactly 1 value. A command such as "> myscript --verbose command" will be
misparsed; the code will assume that "command" is the value of the
flag "--verbose"
}
\examples{
opt_get_command()
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

