% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opt_get_subcommand.R
\name{opt_get_subcommand}
\alias{opt_get_subcommand}
\title{Subcommand to the Rscript.}
\usage{
opt_get_subcommand(opts = commandArgs())
}
\arguments{
\item{opts}{character; Vector from which to parse options
  (default: \code{commandArgs()} )}
}
\value{
character; the subcommand
}
\description{
Return the subcommand of the Rscript, generally the first argument following
\code{--args}.
}
\examples{
opt_get_subcommand()
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

