% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optigrab-package.R
\docType{package}
\name{optigrab-package}
\alias{optigrab}
\alias{optigrab-package}
\title{Command-line parsing desinged for R emphasizing a concise and clean format.}
\description{
The optigrab packages providesa function \code{opt_grab} to retrieve
options/arguments from the command line.  It is useful for running R in
batch mode with \code{R CMD BATCH ...}) or \code{Rscript}.

\emph{optigrab} is similar to the other command-line parsing packages
(\emph{optparse}, \emph{argparse} and \emph{getopt}).  \emph{optigrab} has
two advantages:  It is designed for R in mind and it tries to avoids lengthy,
complex and messy specifications that clutter up the head of programs.  The
emphasis is on a terse and comprehendible syntax that works with R. In most
use cases, the command to retrieve a option is simple and consise:
\code{ f <- opt_grab{ "--flag" } }
\emph{optigrab} allows more powerful options such as coercion, auto --help,
etc. but the most common and simplest use cases do not require the syntax
and complexity of the most complex ones.

\tabular{ll}{
 Package: \tab optigrab\cr
 Type: \tab Package\cr
 Version: \tab 0.1.0\cr
 Date: \tab 2012-12-02\cr
 License: \tab GPLv2\cr
 LazyLoad: \tab no\cr
}
}
\examples{
opts <- c( "--flag", "value" )
 flag <- optigrab:::opt_grab( c("--flag","-f"), opts=opts )  # bar
}
\author{
Christopher Brown Maintainer: Christopher Brown
<chris.brown@decisionpatterns.com>
}
\references{
The Jerk. Dir. Carl Reiner. Perf Steve Martin, Bernadette Peters,
Caitlin Adams. Universal Pictures, 1979.

http://www.gnu.org/prep/standards/standards.html
}
\seealso{
\code{\link[base]{commandArgs}}
}
\keyword{package}

