% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optigrab-package.R
\docType{package}
\name{optigrab}
\alias{optigrab}
\alias{optigrab-package}
\title{Optigrab}
\description{
The optigrab packages providesa function \code{opt_grab} to retrieve 
options/arguments from the command line.  It is useful for running R in 
batch mode with \code{R CMD BATCH ...} or \code{Rscript}.  

GNU-, Java- and Microsoft-style command line options are supported. GNU-style
is the default.

See the \emph{optigrab} vignettes or the github README file for details.
}
\examples{

\dontrun{ 
  opt_get( c("foo","f"))
}

  opts <- c( "--flag", "bar" ) 
  flag <- opt_get( c("foo","f"), opts=opts )  # bar

}
\seealso{
\itemize{
\item \code{\link[=opt_get]{opt_get()}}
\item \code{\link[base:commandArgs]{base::commandArgs()}}
}
}
\keyword{package}
