% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optihelp}
\alias{optihelp}
\title{Provide program usage information and exit.}
\usage{
optihelp(opts = commandArgs())
}
\arguments{
\item{opts}{character.  Vector from which to parse options (default: \code{commandArgs()} )}
}
\description{
Prints program usage information and exits. Usage information comes from
the \code{help} arguments to \code{grab_opt}. Calls to
\code{grab_opt} following the \code{optihelp} call will not be considered, so
it is best to specify make all \code{grab_opt} calls before \code{optihelp}.
}
\examples{
opts <- c( "--foo", "bar")
optigrab:::grab_opt( "--foo", description="Specifies Foo" )
optigrab:::optihelp()
}
\seealso{
\code{\link[base]{commandArgs}}
}
\keyword{utils}

