% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_cl.R
\name{str_to_cl}
\alias{str_to_cl}
\alias{str_to_opts}
\title{Convert a string to \code{cl} (command_line) vector}
\usage{
str_to_cl(x = character())

str_to_opts(...)
}
\arguments{
\item{x}{character; string to parse as if it is a command line 

This is an internal function used predominantly for testing. It might be 
deprecated in the near future.}
}
\value{
A character array that could be similar to that  provided by 
  \code{commandArgs}.
}
\description{
Split a string based on whitespace ignore single- and double quoted entries
}
\note{
not-exported, by design
}
\examples{

  \dontrun{ 
    str <- 'cmd -t "Say Anything" --character \\'Lloyd Dobler\\''
    str_to_cl(str)
    split_ws_nonquote(str)
  }
  
  character(0) \%>\% str_to_cl()  # character(0)
  "" \%>\% str_to_cl()            # character(0)
  "foo" \%>\% str_to_cl()         # "foo"
  "foo bar" \%>\% str_to_cl       # "foo" "bar"
  
  str_to_args( "--foo 'bar baz' qux" )
  
}
\seealso{
\code{\link[base]{commandArgs}} \cr
  \code{\link{cl}}
}

