% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/this_file.R
\name{this_file}
\alias{this_file}
\title{this_file}
\usage{
this_file(opts = commandArgs(), local = TRUE, full.path = TRUE)
}
\arguments{
\item{opts}{character; vector of arguments. (Default: \code{commandArgs()})}

\item{local}{logical; if \code{TRUE} returns the most currently sourced
   script as opposed to the orignal/first source script.
   (Default: \code{TRUE})}

\item{full.path}{logical; Whetther to return the full path to the sourced
   file. (Default: \code{TRUE})}
}
\value{
one-element character vector with the path to the current file;
returns \code{NA} is in an interactive session not in a file.
}
\description{
Name or path to the current file
}
\details{
\code{this_file} returns the name or path of the executing file whehter
the file was invoked from \strong{Rscript} or in an interactive session.
Further it \code{source}

Argument \code{local} controls whether it is the current file (\code{TRUE})
or the orignal, top-level file.
}
\examples{
this_file()
}
\references{
\url{http://stackoverflow.com/questions/1815606/rscript-determine-path-of-the-executing-script}
}
\seealso{
\code{\link{opt_grab}} \cr
  \code{\link{base}{commandArgs}}
}

